# Conflicts:
#	src/Activations/ActivationInterface.php
#	src/Activations/ActivationRepositoryInterface.php
#	src/Activations/EloquentActivation.php
#	src/Activations/IlluminateActivationRepository.php
#	src/Checkpoints/ActivationCheckpoint.php
#	src/Checkpoints/AuthenticatedCheckpoint.php
#	src/Checkpoints/CheckpointInterface.php
#	src/Checkpoints/NotActivatedException.php
#	src/Checkpoints/ThrottleCheckpoint.php
#	src/Checkpoints/ThrottlingException.php
#	src/Cookies/CICookie.php
#	src/Cookies/CookieInterface.php
#	src/Cookies/FuelPHPCookie.php
#	src/Cookies/IlluminateCookie.php
#	src/Cookies/NativeCookie.php
#	src/Hashing/BcryptHasher.php
#	src/Hashing/CallbackHasher.php
#	src/Hashing/Hasher.php
#	src/Hashing/HasherInterface.php
#	src/Hashing/NativeHasher.php
#	src/Hashing/Sha256Hasher.php
#	src/Hashing/WhirlpoolHasher.php
#	src/Laravel/Facades/Activation.php
#	src/Laravel/Facades/Reminder.php
#	src/Laravel/Facades/Sentinel.php
#	src/Laravel/SentinelServiceProvider.php
#	src/Native/ConfigRepository.php
#	src/Native/Facades/Sentinel.php
#	src/Native/SentinelBootstrapper.php
#	src/Permissions/PermissibleInterface.php
#	src/Permissions/PermissibleTrait.php
#	src/Permissions/PermissionsInterface.php
#	src/Permissions/PermissionsTrait.php
#	src/Permissions/StandardPermissions.php
#	src/Permissions/StrictPermissions.php
#	src/Persistences/EloquentPersistence.php
#	src/Persistences/IlluminatePersistenceRepository.php
#	src/Persistences/PersistableInterface.php
#	src/Persistences/PersistenceInterface.php
#	src/Persistences/PersistenceRepositoryInterface.php
#	src/Reminders/EloquentReminder.php
#	src/Reminders/IlluminateReminderRepository.php
#	src/Reminders/ReminderRepositoryInterface.php
#	src/Roles/EloquentRole.php
#	src/Roles/IlluminateRoleRepository.php
#	src/Roles/RoleInterface.php
#	src/Roles/RoleRepositoryInterface.php
#	src/Roles/RoleableInterface.php
#	src/Sentinel.php
#	src/Sessions/CISession.php
#	src/Sessions/FuelPHPSession.php
#	src/Sessions/IlluminateSession.php
#	src/Sessions/NativeSession.php
#	src/Sessions/SessionInterface.php
#	src/Throttling/EloquentThrottle.php
#	src/Throttling/IlluminateThrottleRepository.php
#	src/Throttling/ThrottleRepositoryInterface.php
#	src/Users/EloquentUser.php
#	src/Users/IlluminateUserRepository.php
#	src/Users/UserInterface.php
#	src/Users/UserRepositoryInterface.php
#	src/config/config.php
#	src/migrations/2014_07_02_230147_migration_cartalyst_sentinel.php
#	tests/ActivationCheckpointTest.php
#	tests/BcryptHasherTest.php
#	tests/CICookieTest.php
#	tests/CISessionTest.php
#	tests/CallbackHasherTest.php
#	tests/CheckpointsTest.php
#	tests/EloquentActivationTest.php
#	tests/EloquentPersistenceTest.php
#	tests/EloquentReminderTest.php
#	tests/EloquentRoleTest.php
#	tests/EloquentUserTest.php
#	tests/FuelPHPCookieTest.php
#	tests/FuelPHPSessionTest.php
#	tests/IlluminateActivationRepositoryTest.php
#	tests/IlluminateCookieTest.php
#	tests/IlluminatePersistenceRepositoryTest.php
#	tests/IlluminateReminderRepositoryTest.php
#	tests/IlluminateRoleRepositoryTest.php
#	tests/IlluminateSessionTest.php
#	tests/IlluminateThrottleRepositoryTest.php
#	tests/IlluminateUserRepositoryTest.php
#	tests/NativeCookieTest.php
#	tests/NativeHasherTest.php
#	tests/NativeSessionTest.php
#	tests/PermissibleTraitTest.php
#	tests/PermissionsTraitTest.php
#	tests/SentinelBootstrapperTest.php
#	tests/SentinelTest.php
#	tests/Sha256HasherTest.php
#	tests/StandardPermissionsTest.php
#	tests/StrictPermissionsTest.php
#	tests/ThrottleCheckpointTest.php
#	tests/WhirlpoolHasherTest.php
#	tests/stubs/ci/CI_Input.php
#	tests/stubs/ci/CI_Session.php
#	tests/stubs/fuelphp/Fuel/Core/Cookie.php
#	tests/stubs/fuelphp/Fuel/Core/Session_Driver.php

+++ src/Users/EloquentUser.php
@@ -11,7 +11,7 @@
  * bundled with this package in the LICENSE file.
  *
  * @package    Sentinel
- * @version    2.1.0
+ * @version    2.0.11
  * @author     Cartalyst LLC
  * @license    BSD License (3-clause)
  * @copyright  (c) 2011-2015, Cartalyst LLC
@@ -406,7 +406,9 @@ class EloquentUser extends Model implements RoleableInterface, PermissibleInterf
      */
     public function delete()
     {
-        if ($this->exists) {
+        $isSoftDeleted = array_key_exists('Illuminate\Database\Eloquent\SoftDeletes', class_uses($this));
+
+        if ($this->exists && ! $isSoftDeleted) {
             $this->activations()->delete();
             $this->persistences()->delete();
             $this->reminders()->delete();
